LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/registered-property-ident-function-expected.txt


FAIL The ident() function is resolved in a registered custom property assert_equals: expected "--myident42" but got "none"
